Update crossValidationVSvalidation.py
performed cross validation and validation together to see which one is better to use. even if the error we calculated in cross validation is higher, it's better to use it. Because Cross validation makes sure each data point plays the same percentage of the role in training versus validation. Thus when we look at the variance in the model assessment, there will be less variance in cross validation, yielding more reliable results.